Breaking: new version - addressing financial-modeling API change
