--- .env.example
@@ -1,1 +1,0 @@
-AV_KEY=demo
