+++ .idea/financial-modeling-prep-data-source.time
@@ -265,3 +265,73 @@
 "2020-04-22 19:52:13:293",8000
 "2020-04-22 19:52:17:762",1000
 "2020-04-22 19:52:25:663",6000
+"2020-04-22 19:54:37:931",128000
+"2020-04-22 19:54:59:425",20000
+"2020-04-22 20:01:59:405",5000
+"2020-04-22 20:02:27:948",21000
+"2020-04-22 20:02:31:823",2000
+"2020-04-22 20:03:29:310",29000
+"2020-04-22 20:04:06:487",28000
+"2020-04-22 20:04:20:022",12000
+"2020-04-22 20:04:38:850",10000
+"2020-04-22 20:04:47:485",2000
+"2020-04-22 20:04:51:753",2000
+"2020-04-22 20:05:43:260",25000
+"2020-04-22 20:06:14:707",30000
+"2020-04-22 20:06:21:312",2000
+"2020-04-22 20:06:30:515",7000
+"2020-04-22 20:07:08:735",31000
+"2020-04-22 20:08:49:379",14000
+"2020-04-22 20:08:59:848",1000
+"2020-04-22 20:11:42:489",150000
+"2020-04-22 20:23:49:145",161000
+"2020-04-22 20:27:03:665",166000
+"2020-04-22 20:32:38:245",332000
+"2020-04-22 20:37:37:605",0
+"2020-04-23 00:10:01:428",0
+"2020-04-23 10:46:14:908",107000
+"2020-04-23 10:46:15:905",0
+"2020-04-23 10:49:27:776",175000
+"2020-04-23 10:49:44:741",9000
+"2020-04-23 10:54:07:826",122000
+"2020-04-23 10:55:13:448",49000
+"2020-04-23 10:59:10:047",233000
+"2020-04-23 11:00:25:179",50000
+"2020-04-23 11:00:49:103",22000
+"2020-04-23 11:00:55:725",4000
+"2020-04-23 11:01:12:332",15000
+"2020-04-23 11:05:03:927",230000
+"2020-04-23 11:10:54:111",339000
+"2020-04-23 11:16:21:555",35000
+"2020-04-23 11:16:30:146",4000
+"2020-04-23 11:17:04:629",33000
+"2020-04-23 11:19:40:040",154000
+"2020-04-23 11:21:42:203",103000
+"2020-04-23 11:47:41:450",1000
+"2020-04-23 11:49:32:554",73000
+"2020-04-23 11:50:11:611",10000
+"2020-04-23 11:55:44:356",330000
+"2020-04-23 11:55:55:796",1000
+"2020-04-23 11:57:24:930",85000
+"2020-04-23 11:59:23:345",109000
+"2020-04-23 11:59:29:375",2000
+"2020-04-23 11:59:32:319",1000
+"2020-04-23 12:05:33:048",358000
+"2020-04-23 12:05:57:486",23000
+"2020-04-23 12:08:02:708",14000
+"2020-04-23 12:14:26:175",378000
+"2020-04-23 12:16:32:100",82000
+"2020-04-23 12:19:09:659",138000
+"2020-04-23 12:21:24:393",89000
+"2020-04-23 12:24:44:468",198000
+"2020-04-23 12:24:44:551",0
+"2020-04-23 12:25:23:298",29000
+"2020-04-23 12:25:59:825",7000
+"2020-04-23 12:25:59:933",0
+"2020-04-23 12:30:52:466",285000
+"2020-04-23 12:31:37:974",24000
+"2020-04-23 12:34:48:734",3000
+"2020-04-23 12:36:27:231",2000
+"2020-04-23 12:37:52:168",18000
+"2020-04-23 12:38:56:785",18000
+"2020-04-23 12:42:47:292",2000
