--- src/lib/company/EnterpriseValue.js
@@ -1,7 +1,0 @@
-module.exports = config => {
-	const util = require('./util')(config);
-
-	return {
-		enterprise_value: ({ symbol, period }) => util.fn(['enterprise-value', symbol])({ period })
-	}
-};
