+++ test/cache.test.js
@@ -7,26 +7,21 @@
 /* global performance*/
 
 import FinancialModelingPrepAPI from "../src";
-import * as variables from "../src/mocks/demoVariableSets";
+import * as variables from "./variable_sets";
 
 jest.setTimeout(30000);
 
-const alphaStructure = FinancialModelingPrepAPI.prototype;
-
-describe.each(['data', 'crypto', 'forex', 'performance', 'technical'])("%s", groupKey => {
-	describe.each(Object.keys(alphaStructure[groupKey]))("%s", key => {
-		let varSets = variables[groupKey][key];
-		varSets = Array.isArray(varSets) ? varSets : [varSets];
-		it.each(varSets)("%j", async varSet => {
-			const alpha = new FinancialModelingPrepAPI({ key: 'demo' });
-			const start = performance.now();
-			await alpha[groupKey][key](varSet);
-			const response1time = performance.now();
-			await alpha[groupKey][key](varSet);
-			const response2time = performance.now();
-			console.log(response1time - start + 1,'>',response2time - response1time);
-			expect(response1time - start + 1).toBeGreaterThan(response2time - response1time)
-		});
-	})
+describe.each(Object.keys(variables))("%s", key => {
+	let varSets = variables[key];
+	varSets = Array.isArray(varSets) ? varSets : [varSets];
+	it.each(varSets)("%j", async ({ arg, ...varSet }) => {
+		const fmpAPI = new FinancialModelingPrepAPI();
+		const start = performance.now();
+		await fmpAPI[key](arg, varSet);
+		const response1time = performance.now();
+		await fmpAPI[key](arg, varSet);
+		const response2time = performance.now();
+		console.log(response1time - start + 1, '>', response2time - response1time);
+		expect(response1time - start + 1).toBeGreaterThan(response2time - response1time)
+	});
 });
-
