+++ test/integration.test.js
@@ -5,8 +5,8 @@
 */
 
 import FinancialModelingPrepAPI from "../src";
-import * as variables from "../src/mocks/demoVariableSets";
-import { obtainStructure } from "./utils";
+import * as variables from "./variable_sets";
+import { matchesSnapshot } from "./utils";
 
 jest.setTimeout(30000);
 
@@ -19,11 +19,9 @@ describe.each(Object.keys(variables))("%s", key => {
 	let varSets = variables[key];
 	varSets = Array.isArray(varSets) ? varSets : [varSets];
 	it.each(varSets)("%j", async ({arg, ...varSet}) => {
-		expect.assertions(2);
 		const response = await fmpAPI[key](arg,varSet);
-		const structure = obtainStructure(response);
-		expect(structure).toMatchSnapshot();
-		expect(JSON.stringify(response).length).toBeGreaterThan(20);
+		matchesSnapshot(response);
+		console.log(response)
 	});
 });
